bot-automatisation/requirements.txt
fastapi
uvicorn
python-dotenv
requests
pytest
black
ruff
